More reorg/renaming. Added config check.
